+++ graphql-maven-plugin-logic/src/main/java/com/graphql_java_generator/plugin/conf/CommonConfiguration.java
@@ -5,7 +5,10 @@ package com.graphql_java_generator.plugin.conf;
 
 import java.io.File;
 import java.util.Map;
+import java.util.Objects;
+import java.util.stream.Collectors;
 
+import org.apache.commons.lang.StringUtils;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -140,5 +143,21 @@ public interface CommonConfiguration {
 	 * </UL>
 	 */
 	public boolean isAddRelayConnections();
+
+	/** Logs all the configuration parameters (only when in the debug level) */
+	public void logConfiguration();
+
+	/** Logs all the common configuration parameters (only when in the debug level) */
+	public default void logCommonConfiguration() {
+		getLog().debug("  Common parameters:");
+		getLog().debug("    schemaFileFolder: " + getSchemaFileFolder());
+		getLog().debug("    schemaFilePattern: " + getSchemaFilePattern());
+		getLog().debug("    Templates: "
+				+ (Objects.nonNull(getTemplates())
+						? getTemplates().entrySet().stream()
+								.map(entry -> String.format("%s=%s", entry.getKey(), entry.getValue()))
+								.collect(Collectors.joining(", "))
+						: StringUtils.EMPTY));
+	}
 
 }