+++ graphql-maven-plugin-logic/src/main/java/com/graphql_java_generator/plugin/conf/GenerateClientCodeConfiguration.java
@@ -1,0 +1,15 @@
+/**
+ * 
+ */
+package com.graphql_java_generator.plugin.conf;
+
+/**
+ * This class contains all parameters for the <I>generateClientCode</I> goal/task.
+ * 
+ * @author etienne-sf
+ */
+public interface GenerateClientCodeConfiguration extends GenerateCodeCommonConfiguration {
+
+	// There is currently no parameter that is specific to the client mode
+
+}